+++ src/main.go
@@ -3,16 +3,16 @@ package main
 import (
 	"fmt"
 	"github.com/julienschmidt/httprouter"
-	"github.com/miniwaveme/api/src/config"
+	"github.com/miniwaveme/api/src/conf"
+	"github.com/miniwaveme/api/src/db"
 	"github.com/miniwaveme/api/src/logger"
-	"github.com/miniwaveme/api/src/mongo"
 	"github.com/miniwaveme/api/src/route"
 	"net/http"
 )
 
 func main() {
 
-	conf := config.GetConfig()
+	c := conf.C()
 
 	log := logger.GetLogger()
 	log.Info("application started")
@@ -20,9 +20,8 @@ func main() {
 	r := httprouter.New()
 	route.RegisterRoutesV1(r)
 
-	ds := mongo.Ds
-	mgoDatabase := ds.GetSession().DB(ds.GetDbName())
-	fmt.Println(mgoDatabase.Name)
+	dba := db.DS().DefaultDB()
+	fmt.Println(dba.Name)
 
-	log.Fatal(http.ListenAndServe(":"+conf.GetString("api_port"), r))
+	log.Fatal(http.ListenAndServe(":"+c.GetString("api_port"), r))
 }
